Address code review
- Generalize did_resolver_sov
- Move methods
- Remove did-exchange send_message method
- Address more code review comments
