+++ aries/agents/rust/aries-vcx-agent/src/lib.rs
@@ -9,6 +9,7 @@ extern crate uuid;
 
 mod agent;
 mod error;
+pub mod helper;
 mod http;
 mod services;
 mod storage;
