+++ aries_vcx/src/protocols/did_exchange/state_machine/mod.rs
@@ -6,18 +6,15 @@ pub mod responder;
 
 use std::marker::PhantomData;
 
-use aries_vcx_core::wallet::base_wallet::BaseWallet;
 use chrono::Utc;
-use did_doc_sov::{service::ServiceSov, DidDocumentSov};
+use did_doc_sov::DidDocumentSov;
 pub use helpers::generate_keypair;
 use messages::{
     decorators::{thread::Thread, timing::Timing},
     msg_fields::protocols::did_exchange::problem_report::{
         ProblemCode, ProblemReport, ProblemReportContent, ProblemReportDecorators,
     },
-    AriesMessage,
 };
-use url::Url;
 use uuid::Uuid;
 
 use super::{
@@ -26,11 +23,6 @@ use super::{
         traits::{InvitationId, ThreadId},
     },
     transition::transition_result::TransitionResult,
-};
-use crate::{
-    errors::error::{AriesVcxError, AriesVcxErrorKind, VcxResult},
-    transport::Transport,
-    utils::{encryption_envelope::EncryptionEnvelope, from_did_doc_sov_to_legacy},
 };
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
@@ -127,59 +119,5 @@ impl<I, S> DidExchange<I, S> {
 
     pub fn their_did_doc(&self) -> &DidDocumentSov {
         &self.their_did_document
-    }
-
-    pub fn get_endpoint(&self) -> Option<Url> {
-        let did_document = self.their_did_doc();
-        // todo: no unwrap pls
-        let service = did_document.service().first()?;
-        // todo: will get cleaned up after service is de-generified
-        Some(match service {
-            ServiceSov::Legacy(d) => d.service_endpoint().into(),
-            ServiceSov::AIP1(d) => d.service_endpoint().into(),
-            ServiceSov::DIDCommV1(d) => d.service_endpoint().into(),
-            ServiceSov::DIDCommV2(d) => d.service_endpoint().into(),
-        })
-    }
-    pub async fn encrypt_message(
-        &self,
-        wallet: &impl BaseWallet,
-        message: &AriesMessage,
-    ) -> VcxResult<EncryptionEnvelope> {
-        let sender_verkey = &self
-            .our_did_doc()
-            .resolved_key_agreement()
-            .next()
-            .ok_or_else(|| {
-                AriesVcxError::from_msg(
-                    AriesVcxErrorKind::InvalidState,
-                    "No key agreement method found",
-                )
-            })?
-            .public_key()?
-            .base58();
-        EncryptionEnvelope::create(
-            wallet,
-            json!(message).to_string().as_bytes(),
-            Some(sender_verkey),
-            &from_did_doc_sov_to_legacy(self.their_did_doc().clone())?,
-        )
-        .await
-    }
-
-    pub async fn send_message<T>(
-        &self,
-        wallet: &impl BaseWallet,
-        message: &AriesMessage,
-        transport: &T,
-    ) -> VcxResult<()>
-    where
-        T: Transport,
-    {
-        let msg = self.encrypt_message(wallet, message).await?.0;
-        let service_endpoint = self.get_endpoint().ok_or_else(|| {
-            AriesVcxError::from_msg(AriesVcxErrorKind::InvalidUrl, "No URL in DID Doc")
-        })?;
-        transport.send_message(msg, service_endpoint.clone()).await
     }
 }
