+++ aries_vcx/src/protocols/did_exchange/state_machine/requester/request_sent/helpers.rs
@@ -1,8 +1,12 @@
+use std::sync::Arc;
+
 use aries_vcx_core::ledger::base_ledger::IndyLedgerRead;
 use chrono::Utc;
 use did_doc::schema::verification_method::{VerificationMethod, VerificationMethodType};
 use did_doc_sov::{service::ServiceSov, DidDocumentSov};
 use did_parser::{Did, DidUrl};
+use did_resolver::traits::resolvable::resolution_output::DidResolutionOutput;
+use did_resolver_registry::ResolverRegistry;
 use messages::{
     decorators::{
         thread::{Thread, ThreadGoalCode},
@@ -10,7 +14,7 @@ use messages::{
     },
     msg_fields::protocols::{
         did_exchange::request::{Request, RequestContent, RequestDecorators},
-        out_of_band::invitation::{Invitation as OobInvitation, OobService},
+        out_of_band::invitation::{Invitation as OobInvitation, Invitation, OobService},
     },
 };
 use shared::maybe_known::MaybeKnown;
@@ -18,7 +22,7 @@ use uuid::Uuid;
 
 use crate::{
     common::ledger::transactions::resolve_service,
-    errors::error::{AriesVcxError, AriesVcxErrorKind},
+    errors::error::{AriesVcxError, AriesVcxErrorKind, VcxResult},
     utils::from_legacy_service_to_service_sov,
 };
 
@@ -103,4 +107,61 @@ pub fn construct_request(invitation_id: String, our_did: String) -> Request {
         .content(content)
         .decorators(decorators)
         .build()
+}
+
+pub async fn oob_invitation_to_diddoc(
+    resolver_registry: &Arc<ResolverRegistry>,
+    invitation: Invitation,
+) -> VcxResult<DidDocumentSov> {
+    let mut builder = DidDocumentSov::builder(Default::default());
+
+    let mut resolved_services = vec![];
+    let mut resolved_vms = vec![];
+    let mut resolved_kas = vec![];
+    let mut resolved_dids = vec![];
+
+    for service in invitation.content.services {
+        match service {
+            OobService::SovService(service) => {
+                builder = builder.add_service(service.clone());
+            }
+            OobService::Did(did) => {
+                let parsed_did = Did::parse(did)?;
+                let DidResolutionOutput { did_document, .. } = resolver_registry
+                    .resolve(&parsed_did, &Default::default())
+                    .await?;
+                resolved_services.extend(
+                    did_document
+                        .service()
+                        .iter()
+                        .map(|s| ServiceSov::try_from(s.clone()))
+                        .collect::<Result<Vec<_>, _>>()?,
+                );
+                resolved_vms.extend_from_slice(did_document.verification_method());
+                resolved_kas.extend_from_slice(did_document.key_agreement());
+                resolved_dids.push(parsed_did);
+            }
+            OobService::AriesService(service) => {
+                resolved_services.push(from_legacy_service_to_service_sov(service.clone())?)
+            }
+        }
+    }
+
+    for service in resolved_services {
+        builder = builder.add_service(service);
+    }
+
+    for vm in resolved_vms {
+        builder = builder.add_verification_method(vm.clone());
+    }
+
+    for ka in resolved_kas {
+        builder = builder.add_key_agreement(ka.clone());
+    }
+
+    for did in resolved_dids {
+        builder = builder.add_controller(did);
+    }
+
+    Ok(builder.build())
 }
